// interfaces/prodcuts.ts
export interface Products{
    name: string,
    price: string,
    image: string
}


// export interface req_files{

    // Generated by https://quicktype.io

export interface req_files {
    products?: Product[];

}

export interface Product {
    name:         string;
    data:         Data;
    size:         number;
    encoding:     string;
    tempFilePath: string;
    truncated:    boolean;
    mimetype:     string;
    md5:          string;
}

export interface Data {
    type: string;
    data: any[];
}


export interface Data_image_cloud{
    productId: number
    secure_url: string
    public_id: string
}